Update image_cleanerJustinaSolomon README.md
I created this image cleaning and editing Python software as part of my Thomas College IT team internship and job to crop, re-orient/deskew, and rotate and organize scanned archival library material in preparation for it to become easily tagged with metadata, searchable and usable by researchers. I created it in the span of a month and a half using Python; the program can process, clean, and crop (with an arbitrary border) 30,000+ archival images saving Thomas College 1512 hours of manual labor (or 63 full days).